requirements.txt
streamlit
anthropic
pypdf
reportlab
python-dotenv
pandas
